last changes

+++ app/Http/Services/AuthorService.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Services;
 
+use App\Models\Book;
 use App\Repositories\AuthorRepository;
 use App\Repositories\CategoryRepository;
 
@@ -35,23 +36,30 @@ class AuthorService
     {
         $data = [
             'name' => $parameters['name'],
-            'bio'=>$parameters['bio'],
+            'bio' => $parameters['bio'],
         ];
 
         return $this->authorRepository->createData($data);
     }
 
-    public function updateData($id,array $parameters)
+    public function updateData($id, array $parameters)
     {
         $data = [
             'name' => $parameters['name'],
-            'bio'=>$parameters['bio'],
+            'bio' => $parameters['bio'],
         ];
-       return  $this->authorRepository->updateData($id,$data);
+        return $this->authorRepository->updateData($id, $data);
     }
 
-    public function delete($id){
+    public function delete($id)
+    {
+        $checkBooks = Book::where('author_id', $id)->count();
+        if ($checkBooks > 0) {
+            throw new \Exception('Delete Author Books First');
+        } else {
+            $this->authorRepository->deleteById($id);
+        }
+        return true;
 
-    return $this->authorRepository->deleteById($id);
     }
 }
